+++ test/sweet_tooth/endpoint/module/middleware_test.clj
@@ -11,13 +11,14 @@
 
 (deftest module-groups-middleware-in-correct-order
   (is (= {:duct.core/environment  :test
-          :duct.handler/root      {:middleware [(ig/ref ::dbuddy/authentication)
-                                                (ig/ref ::em/format-response)
-                                                (ig/ref ::em/format-exception)
-                                                (ig/ref ::em/restful-format)
-                                                (ig/ref ::em/merge-params)
-                                                (ig/ref ::null)
-                                                (ig/ref ::em/gzip)]
+          :duct.handler/root      {:middleware [(ig/ref :duct.middleware.buddy/authentication)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/format-response)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/merge-params)
+                                                (ig/ref :sweet-tooth.endpoint.module.middleware-test/null)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/stacktrace-log)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/format-exception)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/restful-format)
+                                                (ig/ref :sweet-tooth.endpoint.middleware/gzip)]
                                    :router     (ig/ref :duct/router)}
           ::dweb/not-found        {:error-handler (ig/ref ::eh/index.html)}
           ::em/format-exception   {:include-data true}
@@ -25,6 +26,7 @@
           ::em/restful-format     {:formats [:transit-json]}
           ::em/merge-params       {}
           ::em/format-response    {}
+          ::em/stacktrace-log     {}
           ::dbuddy/authentication {:backend :session}
           ::eh/index.html         {}}
          (duct/prep-config {:duct.profile/base                      {:duct.handler/root     {:middleware [(ig/ref ::null)]}
@@ -33,17 +35,19 @@
 
 (deftest exclude-middleware
   (is (= {:duct.core/environment :test
-          :duct.handler/root     {:middleware [(ig/ref ::em/format-response)
-                                               (ig/ref ::em/format-exception)
-                                               (ig/ref ::em/merge-params)
-                                               (ig/ref ::null)
-                                               (ig/ref ::em/gzip)]
+          :duct.handler/root     {:middleware [(ig/ref :sweet-tooth.endpoint.middleware/format-response)
+                                               (ig/ref :sweet-tooth.endpoint.middleware/merge-params)
+                                               (ig/ref :sweet-tooth.endpoint.module.middleware-test/null)
+                                               (ig/ref :sweet-tooth.endpoint.middleware/stacktrace-log)
+                                               (ig/ref :sweet-tooth.endpoint.middleware/format-exception)
+                                               (ig/ref :sweet-tooth.endpoint.middleware/gzip)]
                                   :router     (ig/ref :duct/router)}
           ::dweb/not-found       {:error-handler (ig/ref ::eh/index.html)}
           ::em/format-exception  {:include-data true}
           ::em/gzip              {}
           ::em/merge-params      {}
           ::em/format-response   {}
+          ::em/stacktrace-log    {}
           ::eh/index.html        {}}
          (duct/prep-config {:duct.profile/base                      {:duct.handler/root     {:middleware [(ig/ref ::null)]}
                                                                      :duct.core/environment :test}